commons-mongo: - updated driver to 3.5.0 - made specific driver depenencies optional - moved driver-related utilities to separate packages
